requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
docusign-esign==4.0.0
Flask==2.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-OAuthlib==0.9.6
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
ndg-httpsclient==0.5.1
nose==1.3.7
oauthlib==2.1.0
paypalrestsdk==1.13.3
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
retrying==1.3.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.0.3
